Crate any_ascii

Unicode to ASCII transliteration.

Converts Unicode text to a reasonable representation using only ASCII.

For most characters in Unicode, AnyAscii provides an ASCII-only replacement string. Text is converted character-by-character without considering the context. The mappings for each script are based on popular existing romanization schemes. Symbolic characters are converted based on their meaning or appearance. All ASCII characters in the input are left unchanged, every other character is replaced with printable ASCII characters. Unknown characters are removed.

This crate supports no_std + alloc.

Functions

any_ascii

Transliterates a Unicode String into ASCII.

any_ascii_char

Transliterates a Unicode char into ASCII.